py_requirements.txt
appdirs==1.4.4
asgiref==3.2.10
autopep8==1.5.4
cffi==1.14.2
cryptography==3.1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1
django-cors-headers==3.4.0
django-heroku==0.3.1
django-rest-framework==0.1.0
django-rest-knox==4.1.0
djangorestframework==3.11.1
filelock==3.0.12
gunicorn==20.0.4
importlib-metadata==1.7.0
importlib-resources==3.0.0
parameterized==0.7.4
pbr==5.4.5
psycopg2==2.8.5
pycodestyle==2.6.0
pycparser==2.20
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
stevedore==3.2.0
toml==0.10.1
virtualenv==20.0.31
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
whitenoise==5.2.0
zipp==3.1.0
